Massive changes...
Added level system and fixed handler bug where the game wouldn't start
sometimes because of handler not running through objects. Added menu so
that separate parts of the game can be accessed. Added coins, and
removed sound play back so for efficiency in debugging, also improved
sticky keys for player movement. Added a couple files to the resources
folder, still trying to get that to work. Code is fully maluable, use
wisely :)